requirements.txt
streamlit==1.31.1
streamlit-option-menu==0.3.12
google-generativeai==0.4.0
pillow==10.2.0
deep-translator
gTTS==2.5.2
playsound==1.3.0